backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
feedparser
beautifulsoup4
apscheduler
python-dotenv
sqlalchemy
aiosqlite
